requirements.txt
streamlit
PyPDF2
faiss-cpu
langchain
sentence-transformers
google-generativeai
python-dotenv
